requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy
Pillow
clip-by-openai
tqdm
pycocotools
pandas
requests
ftfy
jupyter
matplotlib
nltk
omegaconf
opencv-python
pytest
regex
scikit-learn
scipy
scikit-image
torchray
wandb
spacy

